% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tribF.R
\name{trib.CoF}
\alias{trib.CoF}
\title{\code{trib.CoF} gets file name length and returns coeficient of friction in
bar plot generates global var CF}
\usage{
trib.CoF(x, DATAOUT = FALSE)
}
\arguments{
\item{x}{Numeric value length of file name for truncation. Truncated file
name is given at output for ease of use.}

\item{DATAOUT}{logical scalar checking whether give data out or not}
}
\value{
Plots a bar plot of Coeficient of Friction generated from all files
  in working directory and generates a vector called CF that contains Average
  Coefficient of Friction values

  Set work directory to file containing example data for Tribometer directory
  should be
}
\description{
\code{trib.CoF} gets file name length and returns coeficient of friction in
bar plot generates global var CF
}
\section{Warning}{
 Before using this function please set your working
  directory to the directory that contains .csv files from tribometer. File
  structure should be as follows
  \tabular{ccccc}{

  Time  \tab  Distance   \tab  Laps  \tab  mu  \tab   FrictionForce \cr
  0,100 \tab 0     \tab      0,00000  \tab 0,18478 \tab      0,55435
  }

  You can check example datasets at package files exdata/tribometer
}
\examples{
setwd(paste(path.package("trib"),"/exdata/tribometer",sep=""))
trib.CoF(2)
tribDATA<-trib.CoF(2,DATAOUT = TRUE)
}

