% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tribF.R
\name{trib.FF}
\alias{trib.FF}
\title{\code{trib.FF} gets file name length, filterValue and Friction vs Distance
combined graph and/or single graphs for data}
\usage{
trib.FF(x, filterValue, CombinedGraph = TRUE, SingleGraph = TRUE)
}
\arguments{
\item{x}{Numeric value length of file name for truncation. Truncated file
name is given at output for ease of use.}

\item{filterValue}{Numeric value for setting for rolling average from package
\code{zoo}}

\item{CombinedGraph}{logical scalar checking whether to plot combined graph
of all data or not}

\item{SingleGraph}{logical scalar checking whether to plot single graphs of
  all data or not}
}
\value{
Plots Friction vs Distance scatter plots with given filter value if
  CombinedGraph and or SingleGraph options are set, else will return nothing

Set work directory to file containing example data
for Tribometer directory should be
}
\description{
\code{trib.FF} gets file name length, filterValue and Friction vs Distance
combined graph and/or single graphs for data
}
\section{Warning}{
 Before using this function please set your working
  directory to the directory that contains .csv files from tribometer. File
  structure should be as follows
  \tabular{ccccc}{

  Time  \tab  Distance   \tab  Laps  \tab  mu  \tab   FrictionForce \cr
  0,100 \tab 0     \tab      0,00000  \tab 0,18478 \tab      0,55435
  }

  You can check example datasets at package files exdata/tribometer
}
\examples{
setwd(paste(path.package("trib"),"/exdata/tribometer",sep=""))

trib.FF(2,100)

trib.FF(2, 100, CombinedGraph=FALSE)

trib.FF(2, 100, SingleGraph=FALSE)
}

