% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tribF.R
\name{trib.names}
\alias{trib.names}
\title{\code{trib.names} returns the names of .csv
files in working directory and stores table values to a global variabla All,
filenames to global variabla filenames and returns truncated names of files.}
\usage{
trib.names(x, DataFF = TRUE, DEncode = "UTF-8")
}
\arguments{
\item{x}{Numeric value length of file name for truncation. Truncated file
name is given at output for ease of use.}

\item{DataFF}{logical scalar, checking whether to use CSV style for
Tribometer of Profilometer}

\item{DEncode}{is a string that specifies file encoding for import}
}
\value{
returns truncated file names, also generates list All and filenames
  vector
}
\description{
\code{trib.names} returns the names of .csv
files in working directory and stores table values to a global variabla All,
filenames to global variabla filenames and returns truncated names of files.
}
\section{Warning}{

If wrong type is used at DataFF \code{read.csv} will have problems
}
\examples{
setwd(paste(path.package("trib"),"/exdata/tribometer",sep=""))

trib.names(2)

setwd(paste(path.package("trib"),"/exdata/profile",sep=""))

trib.names(2,DataFF=FALSE)
}

