% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trib.R
\docType{package}
\name{trib}
\alias{trib}
\alias{trib-package}
\title{Tribology Measurement Analysis and Visualization}
\description{
trib provides three important functions: CoF, FF and profile. The
function names is used for file import.
}
\section{trib Functions}{


  The trib functions are as follows

  \code{trib.CoF} calculates Average Coefficient of Fricton from tribometer
  data ,

  \code{trib.FF} plot Friction Force vs Distance plots from tribometer data


  \code{trib.profile} analyze surface profilometer data
}

